fix: move para files from the '/learning' folder to the root. Delete old folder
